Move SideBar component to typescript
