+++ package.json
@@ -48,6 +48,7 @@
     "@types/react-router-dom": "5.1.3",
     "@types/redux-actions": "2.6.1",
     "@types/styled-components": "4.4.0",
+    "@types/webpack-env": "1.14.1",
     "@typescript-eslint/eslint-plugin": "2.10.0",
     "@typescript-eslint/parser": "2.10.0",
     "awesome-typescript-loader": "5.2.1",
