+++ src/index.tsx
@@ -21,4 +21,6 @@ render(
   document.getElementById('app')
 );
 
-// module.hot.accept();
+if ((module as any).hot) {
+  (module as any).hot.accept()
+}
